requirements.txt
pillow>=9.0.0
numpy>=1.21.0
cryptography>=3.4.0
click>=8.0.0